Fixes and added functions
Changed the way that the year variable is handled, now the  year is always stored in it's printing format instead of being a number after 1900.
Also added functions setLocation and setMove.
Also deleted old comments and testing variables that are not needed.
If everything goes well this should be the last changes to this class this iteration.
All work done by Bryce Brooks